45020 - PAYMENTSESSION : Compute the sessionTotalAmount of the payment session
