+++ axelor-account/src/main/java/com/axelor/apps/account/service/PaymentSessionService.java
@@ -26,4 +26,6 @@ public interface PaymentSessionService {
   public void setBankDetails(PaymentSession paymentSession);
 
   public void setJournal(PaymentSession paymentSession);
+
+  public void computeTotalPaymentSession(PaymentSession paymentSession);
 }
