+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/InvoiceTermService.java
@@ -218,6 +218,4 @@ public interface InvoiceTermService {
 
   public Integer massRefusePfp(
       List<Long> invoiceTermIds, CancelReason reasonOfRefusalToPay, String reasonOfRefusalToPayStr);
-
-  public void computeTotalPaymentSession(PaymentSession paymentSession);
 }
