+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/InvoiceTermServiceImpl.java
@@ -29,7 +29,7 @@ import com.axelor.apps.account.db.PfpPartialReason;
 import com.axelor.apps.account.db.SubstitutePfpValidator;
 import com.axelor.apps.account.db.repo.InvoiceRepository;
 import com.axelor.apps.account.db.repo.InvoiceTermRepository;
-import com.axelor.apps.account.db.repo.PaymentSessionRepository;
+import com.axelor.apps.account.service.PaymentSessionService;
 import com.axelor.apps.account.service.app.AppAccountService;
 import com.axelor.apps.account.service.config.AccountConfigService;
 import com.axelor.apps.base.db.CancelReason;
@@ -553,7 +553,7 @@ public class InvoiceTermServiceImpl implements InvoiceTermService {
           fillEligibleTerm(paymentSession, invoiceTerm);
           invoiceTermRepo.save(invoiceTerm);
         });
-    computeTotalPaymentSession(paymentSession);
+    Beans.get(PaymentSessionService.class).computeTotalPaymentSession(paymentSession);
   }
 
   private String retrieveEligibleTermsQuery() {
@@ -873,21 +873,5 @@ public class InvoiceTermServiceImpl implements InvoiceTermService {
       managePassedForPayment(invoiceTerm);
       invoiceTermRepo.save(invoiceTerm);
     }
-  }
-
-  @Override
-  @Transactional
-  public void computeTotalPaymentSession(PaymentSession paymentSession) {
-    BigDecimal sessionTotalAmount =
-        invoiceTermRepo
-            .all()
-            .filter(
-                "self.paymentSession = :paymentSession AND self.isSelectedOnPaymentSession = TRUE")
-            .bind("paymentSession", paymentSession.getId())
-            .fetchStream()
-            .map(InvoiceTerm::getAmountPaid)
-            .reduce(BigDecimal.ZERO, BigDecimal::add);
-    paymentSession.setSessionTotalAmount(sessionTotalAmount);
-    Beans.get(PaymentSessionRepository.class).save(paymentSession);
   }
 }
