+++ axelor-account/src/main/java/com/axelor/apps/account/web/InvoiceTermController.java
@@ -26,6 +26,7 @@ import com.axelor.apps.account.db.repo.AccountInvoiceTermRepository;
 import com.axelor.apps.account.db.repo.InvoiceTermRepository;
 import com.axelor.apps.account.db.repo.PaymentSessionRepository;
 import com.axelor.apps.account.exception.IExceptionMessage;
+import com.axelor.apps.account.service.PaymentSessionService;
 import com.axelor.apps.account.service.invoice.InvoiceService;
 import com.axelor.apps.account.service.invoice.InvoiceTermService;
 import com.axelor.apps.base.service.app.AppBaseService;
@@ -348,7 +349,7 @@ public class InvoiceTermController {
   public void computeTotalPaymentSession(ActionRequest request, ActionResponse response) {
     try {
       InvoiceTerm invoiceTerm = request.getContext().asType(InvoiceTerm.class);
-      Beans.get(InvoiceTermService.class)
+      Beans.get(PaymentSessionService.class)
           .computeTotalPaymentSession(invoiceTerm.getPaymentSession());
       response.setReload(true);
     } catch (Exception e) {
