+++ axelor-account/src/main/resources/views/Charts.xml
@@ -1047,7 +1047,7 @@
     title="Custom">
     <dataset type="jpql"><![CDATA[
 
-	 select CONCAT(ceil(sum(self.amountRemaining)), ' €') as Amount, _partner.name as Supplier, _bankDetails.fullName as BankDetails
+	 select CONCAT(round(sum(self.amountPaid),2), ' €') as Amount, _partner.name as Supplier, _bankDetails.fullName as BankDetails
 	 from InvoiceTerm as self
 	 left join self.invoice.partner as _partner
 	 left join self.bankDetails as _bankDetails
